More Data and Message Tweaks

+++ datamodels/gyroscopic.go
@@ -2,11 +2,11 @@ package datamodels
 
 // This struct represents a single gyroscope sensor reading
 type GyroscopeReading struct {
-	EventState            int `json:"EventState"`
-	XAxisAngle            int `json:"XAxisAngle"`
-	YAxisAngle            int `json:"YAxisAngle"`
-	ZAxisAngle            int `json:"ZAxisAngle"`
-	GyroscopicSensorState int `json:"GyroSensorState"`
+	EventState            int     `json:"EventState"`
+	XAxisAngle            float64 `json:"XAxisAngle"`
+	YAxisAngle            float64 `json:"YAxisAngle"`
+	ZAxisAngle            float64 `json:"ZAxisAngle"`
+	GyroscopicSensorState int     `json:"GyroSensorState"`
 }
 
 // This struct represents a single gyroscope event
